pyproject.toml
[project]
name = "cement-ai-backend"
version = "1.0.0"
description = "Google Cloud-powered Cement Plant AI Optimization Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.1",
    "google-cloud-aiplatform>=1.70.0",
    "cloud-sql-python-connector>=1.11.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-vision>=3.8.0",
    "google-cloud-logging>=3.11.0",
    "google-generativeai>=0.8.3",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-google-vertexai>=2.0.10",
    "langgraph>=0.2.45",
    "sqlalchemy>=2.0.35",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.12",
    "websockets>=13.1",
    "pandas>=2.2.0",
    "numpy>=2.1.0",
    "pillow>=11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
